More bugfixes
Added the missing functionality for the fullScreen toggle button.

Fixed the Alarm getting stuck playing when resetting timer or just toggling the volume button.

Reduced timer and stopwatch intervals. Updating the timers every 10ms is too often as it causes too much re-rendering of the components for the browser to handle. This causes the timer run slowly. I've increased the interval time to address this which has also fixed the progress bar.